dev-requirements.txt
django-debug-toolbar==1.0.1
ipdb==0.8
ipython==1.2.0
pip-tools==0.3.4
pytest-django==2.6
git+git://github.com/mverteuil/pytest-ipdb.git#egg=pytest-ipdb
